refactor. 유틸 테스트 클래스 분할
